Moved sources into src folder
Remove clutter in root of repo.
Updated paths in tests etc.
